Demo CSV Importer

+++ src/main/java/com/vaistra/repositories/DemoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DemoRepository extends JpaRepository<DemoCSV,Integer> {
+public interface DemoRepository extends JpaRepository<DemoCSV,Long> {
 
 }
